requirements.txt
django>=4.0.0,<4.1.0
django-filter>=21.1
django-guardian>=2.4.0
djangorestframework>=3.13.0
#psycopg2>=2.9.0
mysqlclient
pyYAML>=6.0
#djangorestframework-gis
